Add a composefs Rust crate
Right now this includes just a parser/generator for dump files,
which is helpful for e.g. a flow like taking a tarball, parsing
it in Rust, and then serializing the entry to a composefs dumpfile line.

I'd like to add more opinionated and advanced functionality
related to composefs implemented in this crate - *especially* integration
with containers/OCI.  My thought is that in the short term we can
publish this crate without having the core
C code depend on it.

I plan to use this crate in bootc.

However, more medium term, it may make sense to offer a CLI
tool from this crate, in which case there will be some tension with
the CLI tools that happen to be implemented in C, as commonly
people would

Signed-off-by: Colin Walters <walters@verbum.org>