src: add selinux labelling support
This should be a fairly straight-forward operation on the in-memory
filesystem tree, but there are a couple of complications:

First: the selabel APIs want the policy to be somewhere on the host
filesystem so it can open and mmap the files.  Fortunately this is easy
to overcome: we can create a temporary directory and extract the policy
to it.

Second: our use of Rc<Leaf> to handle hardlinks is causing problems
here.  We know that we're the only code trying to mutate the filesystem
tree when we run the label process, but we can't possibly convince Rust
about this.  This is why Rc<RefCell<T>> is so popular but I wonder if
there's another way to avoid this....

Third: speaking of hardlinks: what policy are we supposed to apply when
two different filenames for the same hardlink end up telling us two
different labels to use?

Otherwise, this seems to be working well so far.  Still very much WIP
though.

Closes #29

Signed-off-by: Allison Karlitskaya <allison.karlitskaya@redhat.com>