fsverity: Use thiserror for measurement, add a compare API
Change the API to use thiserror instead of anyhow, as we
want in some cases more precise error mappings.

Also by default change the API to return `Option`
for callers that want to not have missing verity be a fatal
error.

Finally, add a high level compare API that most of the callers
really wanted.

Note I didn't try to add a wrapper error enum for the
verity *enablement* part yet.

Signed-off-by: Colin Walters <walters@verbum.org>