src: add internal erofs writer code
With this change, we create the erofs filesystem for ourselves, instead
of using the external mkcomposefs CLI.

This currently produces a different output than the output of
mkcomposefs.  We're planning to make changes here to make the output
more similar to mkcomposefs while pursuing changes in mkcomposefs to
make it more similar to our format.  This work is being tracked as an
issue in https://github.com/containers/composefs/issues/198 and a first
pull request in https://github.com/containers/composefs/pull/410 .

Closes #56

Signed-off-by: Allison Karlitskaya <allison.karlitskaya@redhat.com>