tests: use `test_with` and skip tests that need `mkcomposefs`
This commit adds the `test_with` crate and skips the tests that
require `mkcomposefs`.

It uses use the minimal configuration of `tests_with` and with that
we now pull in 229 `cargo tree` lines vs the pervious 217 (fwiw,
the default configuration of tests_with incrases the crates from
217 to 497). Even the extra 12 might be too much for this tiny
feature, feel free to close again if you feel this is not worth it.

Signed-off-by: Michael Vogt <mvogt@redhat.com>