fs: track all files as potential hardlinks
While pairing with Timothée we discovered that the examples don't work
on his system because the `cfsctl create-image` on the inside of the
container build fails to detect all hardlinks.  The difference is that
his podman storage driver is using fuse-overlayfs, which doesn't always
accurately report `st_nlink`.  The behaviour is odd: when observing the
hardlink via the first filename it will report a given `(st_dev,
st_ino)` with `st_nlink` of 1 but when inspecting it via the second
filename it will report the same `(st_dev, st_ino)` but with `st_nlink`
of 2 this time.

Let's just be "less clever" and store all non-directory inodes in our
hashtable as potential hardlinks.

See https://github.com/containers/fuse-overlayfs/issues/435

Signed-off-by: Allison Karlitskaya <allison.karlitskaya@redhat.com>
Co-Authored-By: Timothée Ravier <tim@siosm.fr>