src: refactor mount fs-verity handling
Add a common function in Repository to open images, checking fs-verity
if appropriate.  Use it for both mount() and pivot_sysroot().

Pass the name of the image through to the mount code so that it can be
given as the "source" of the overlay instead of none.  We use a string
in the format "composefs:[fs-verity digest]".  This shows up in
/proc/mounts, findmnt, etc. and is useful for knowing exactly which
image is mounted (although it should be treated as strictly
informational and not be relied on for security purposes).

Closes #65

Signed-off-by: Allison Karlitskaya <allison.karlitskaya@redhat.com>