fsverity: introduce EnableVerifyError::FilesystemNotSupported error
This commit improves the UX of the error reporting when enabling
fs-verity fails because the underlying filesystem does not support
this. Here we currently just error with:
```
Inappropriate ioctl for device"
```
with the new code we error with:
```
Filesystem not supported by fs-verity
```

Note that the test relies on the fact that /dev/shm is not supported
by fs-verity.

Closes: https://github.com/containers/composefs-rs/issues/59
Signed-off-by: Michael Vogt <michael.vogt@gmail.com>