examples/bls: enable ssh server
We need to generate the ssh keys as part of the container build because
we don't have a working writable /etc on the booted system yet.

Add the cockpit test key to root's .ssh/authorized_keys.  We're going to
be running tests on the images using Cockpit's testing framework soon.

Signed-off-by: Allison Karlitskaya <allison.karlitskaya@redhat.com>