examples/bls: more similarity to Cockpit VMs
Use ttyS0 as our serial console and make sure our image name ends with
-efi (which is how Cockpit's VM tools know how to boot us in EFI mode).
Make the name more descriptive as well, since this is visible in `virsh
list` output as well as assigned via DHCP as part of the machine's
hostname.

Drop the run symlink from the examples/bls directory: this would display
a graphics console, which is useless now that the console is on the
serial port.

It's now possible to boot the built image using Cockpit's `vm-run`
command and connect to it via the usual `ssh c` alias.

Signed-off-by: Allison Karlitskaya <allison.karlitskaya@redhat.com>