examples: move away from fix-verity script
Add a check-config script that verifies that the user has some of the
required tooling to build the image (before attempting to do so) and
makes sure that their fs-verity setup is sufficient (in terms of
fs-verity support on the working directory, plus capable tools).

The fix-verity script is now disabled by default, which means that
check-config will be checking that capable versions of systemd-repart
and mkfs.ext4 are installed.  Those versions are not packaged anywhere
(or even merged upstream yet) so we have a script to help install them.

You can also use the FIX_VERITY=1 environment variable to get the old
behaviour back.

Signed-off-by: Allison Karlitskaya <allison.karlitskaya@redhat.com>