examples/bls: use mkinitcpio for initramfs generation
Arch Linux uses dracut-ng which by default runs in `hostonly` mode with
no easy way to override this behaviour via kernel-install. `hostonly`
mode misses virt_blk, ext4 and other kernel modules we would need to
pivot to composefs before `switch_root`.

The default initramfs generator for Arch Linux is `mkinitcpio` which by
default is without `systemd` and the root filesystem is mounted on
`/new_root` instead of `/sysroot`.

Since there is no systemd in the initramfs the root partition has to be
hardcoded in the kernel command line and custom composefs hooks to call
composefs-pivot-sysroot. The hooks are GPL2 so they can eventually be
upstreamed into `mkinitcpio` itself.

Signed-off-by: Jelle van der Waa <jvanderw@redhat.com>