mount: start working with fds instead of paths
Kernel 6.15 is going to get support for dealing with floating
filesystems as overlayfs layers which gives us a much cleaner way to
build our complicated world (which is only going to get more complicated
as we start adding features to composefs-pivot-sysroot).

Move things over to being based more around fds and less around paths.
In particular, change the mount code so that it works purely in terms of
fds and the the use of a temporary mountpoint is treated as a temporary
compatibility shim.

Change `Repository::object_path()` currently returning a `Path` to
`Repository::object_dir()` returning an fd and use that as our datadir
when mounting.  This was the only thing that required the `path`
property on the repository, so we can drop that now.

In terms of what's happening there's not a whole lot of change here
(other than some extra indirection) but this lays the groundwork for
rewriting the pivot code in a much cleaner style.

Signed-off-by: Allison Karlitskaya <allison.karlitskaya@redhat.com>