examples: make 'os' variant more explicit
Instead of selecting the Containerfile by way of passing a `-f` flag to
the various examples/*/build scripts, make those scripts accept an OS
name as their only parameter.  Use this to select the Containerfile and
also customize the name of the image according to it.

Update the test runner and GitHub Actions workflows accordingly.

Signed-off-by: Allison Karlitskaya <allison.karlitskaya@redhat.com>