examples/unified-secureboot: run under CI
Update the unified-secureboot example to work with the Cockpit bots
libraries in the same way as the other examples do.  Add it to CI.

When run inside of CI we don't actually enable secure boot (because
Cockpit bots doesn't support it) but it will still be checked when we
use the ./run script.  Now that this example is the only one still using
that script, update it for serial console support and move it from the
examples/common to the unified-secureboot directory.

Fixes #83

Signed-off-by: Allison Karlitskaya <allison.karlitskaya@redhat.com>