mount: more fd-centric cleanups
We now use the system mount APIs in a more modern way: the filesystem
tree is now assembled purely from file descriptors and mounted in place
only after it's complete, resulting in more readable code.  This depends
on a very new kernel: the merge window on 6.15 isn't closed yet, but we
already depend on many of the feature of the mount API that got added in
this release.

In particular, we make use of these features, with fallbacks:
 - the ability to use a detached mount as an overlayfs layer
 - the ability to supply O_PATH fds as overlayfs layers

At the same time, we preserve backwards compatibility to older kernels
via a compatibility layer which remains mostly isolated in a separate
file.  We even add support for creating loopback devices for
compatibility with RHEL 9 (which can't mount erofs from a file).  The
inclusion of the compatibility code is controlled by the feature flags
`pre-6.15` and `rhel9` (which implies `pre-6.15`).

When running on pre-6.15 kernels (which is enabled by default), very
little has actually changed in terms of what's happening at the syscall
level.

Signed-off-by: Allison Karlitskaya <allison.karlitskaya@redhat.com>