src/bin: new composefs-setup-root binary
Take a copy of the existing composefs-pivot-sysroot command and
substantially rewrite it.

The new command understands ostree's /usr/lib/ostree/prepare-root.conf
file (transient overlays for /etc and /, also adding support /var).  In
addition it's possible to specify that /etc and /var are one of:
  - none: no mount, will be readonly contents of composefs at runtime
  - bind: straight bind-mount from the state directory
  - overlay: state directory contains the upperdir of an overlay
  - transient: alias for transient=true (ie: overlay with tmpfs)

This follows the /sysroot/state/ layout discussed in #38.

The default for /etc is 'overlay' and the default for /var is 'bind'.

In general the new command focuses less on absolute minimalism: we now
have proper commandline parsing and our config file is parsed as toml
via serde.  This makes the command (which gets included in the
initramfs) a fair bit bigger: it's 1.2MB now (but compresses to about
half that).  We can deal with that later if it's really a problem,
though.

We also embrace the "build the tree purely from fds" approach from the
previous commit and even mount the various subdirectories (var, etc.)
into the still-detached composefs mount object.  This also depends on a
new feature in 6.15 (mounting into detached mounts) which means we need
to adjust the order of operations for older kernels (to mount the new
root directory first) but this is conditionalized with 'pre-6.15'.

Signed-off-by: Allison Karlitskaya <allison.karlitskaya@redhat.com>