examples: use composefs-setup-root
Move the examples to make use of the new `composefs-setup-root` command.
This is mostly an application of renaming and `sed` but we also have to
adjust the directories created on the resulting system image: they need
to have a `state/` directory with a `etc/` and `var/` present inside of
the deployment.

Signed-off-by: Allison Karlitskaya <allison.karlitskaya@redhat.com>