tests: add an new composefs-setup-root test
This script can be run under `unshare -Umr` to test that
composefs-setup-root is working properly, without building a VM image.

This helped a lot with manual testing during the original development of
composefs-setup-root and will be expanded in the future to include more
scenarios (different types of /etc and /var mounting, transient
overlays, and so on).

Signed-off-by: Allison Karlitskaya <allison.karlitskaya@redhat.com>