oci: add support for pulling decompressed tar
Check the media_type on the descriptor before attempting gzip
decompression.  If the media type indicates an uncompressed layer, then
consume the stream directly.  Add zstd support as well.

Unfortunately the code is a bit awkward: split_async() is generic, so we
need to make a separate call for each potential underlying AsyncBufRead
impl.

Signed-off-by: Allison Karlitskaya <allison.karlitskaya@redhat.com>