image: wrap LeafContent::Regular{Inline,External}
Create a new LeafContent::Regular which covers both cases and contains
another enum which can be either RegularFile::Inline or ::External.

This is a bit more typing in some cases but lets us consolidate a couple
of match statements and also introduces the possibility of APIs which
can access regular files in uniform ways.

Signed-off-by: Allison Karlitskaya <allison.karlitskaya@redhat.com>