image: change inline data from Vec<u8> to Box<[u8]>
This makes more sense, and we don't risk accidentally carrying around
spare capacity.  Do the same for splitstream as well.  This requires a
couple of adjustments around the tree, but nothing major.

Take the chance to make better use of the new Buffer API in rustix to
avoid a copy (from a slice subset of a vector back into a vector) when
reading the file data from the filesystem.

Signed-off-by: Allison Karlitskaya <allison.karlitskaya@redhat.com>