uki: new file for querying UKI boot labels
This contains a single function to query an appropriate boot label for a
UKI image, which might be interesting for supporting bootloaders which
don't implement the bootloader specification and is also likely to be
used for showing installed bootloader entries in the future.

Nothing in the crate uses this right now, but it's a public API.

This is parsing code, so add a bunch of unit tests for various bits of
the process.  We could easily have more, and this would probably benefit
from fuzzing as well, but that's enough for now.  I've also tested it
manually with a couple of UKIs and also with systemd-bootx64.efi and it
produces reasonable results, but I don't want to include those large
files in git.

Signed-off-by: Allison Karlitskaya <allison.karlitskaya@redhat.com>