src/fsverity: support SHA-512 and block sizes
Use some generics magic to add support for SHA-512 and different block
sizes in our fs-verity digest code.

Expose the hash algorithm selection via the public API but not the block
size: it's not possible to specify default values for the generic
parameters on function calls, and although the choice of algorithm can
always be infered by the return type from all of our callers, there is
no information that would help determine the default block size.

Signed-off-by: Allison Karlitskaya <allison.karlitskaya@redhat.com>