src/fsverity: clean up error enums
Stop leaking rustix out via our EnableVerityError enum: we pay for this
by losing PartialEq, but this bring us in line with the very deliberate
decision to use std::io::Error that we made for MeasureVerityError.

Add a separate enum member for the "fs-verity already enabled" case.

Update the tests accordingly.

Signed-off-by: Allison Karlitskaya <allison.karlitskaya@redhat.com>