Cargo.toml: list actual minimum dependencies
In 72d8b71465c6 ("Cargo.toml: update all the things") we randomly raised
all of our dependencies to the latest versions, mostly because I don't
know what I'm doing when it comes to these things.  That was a mistake:
trying to use the latest version of composefs-rs from bootc causes all
kinds of dependency conflicts now.

So let's try and do the opposite and try to list the minimum version
that:
 - is semver compatible with the latest version (so we get updates)
 - doesn't break

I tried playing with automated tooling for this but it wasn't easy to
get it working without a nightly toolchain (which I don't have setup),
and I was also interested in exploring the actual reasons for the
requirements, so I went with a simple manual process: I dropped all of
the version numbers in [dependencies] (but not [dev-dependencies]) to
their lowest semver-compatible version and added "=" in front.

Then I adjusted upwards for yanked releases:
 - sha2 0.10.0 -> 0.10.1
 - tempfile 3.0.0 -> 3.0.2
 - tokio 1.0.0 -> 1.0.1
 - xxhash-rust 0.8.0 -> 0.8.2

Then (because of '=' dependencies) I needed to make adjustments for
conflicts with dependencies specified by other packages:
 - log 0.4.0 -> log 0.4.8 to satisfy an env_logger dependency
 - regex-automata 0.4.0 -> 0.4.4 for regex (via oci-spec)
 - serde 1.0.0 -> 1.0.145 for toml
 - tokio 1.0.1 -> 1.24.2 for async-compression

Then I actually tried to build and run unit tests to determine which
versions were actually required in this crate, and bumped some more
things:
 - anyhow 1.0.0 -> 1.0.87 to support the automatic conversions we
   require via use of `?`
 - clap 4.0.0 -> 4.0.1 was the first version with #[clap(version)]
 - tar 0.4.0 -> 0.4.38 for PaxExtensions::new()
 - tempfile 3.0.2 -> 3.8.0 for tempfile_in()

After that, I ran the examples and everything there seemed to work
properly without additional changes, so I dropped the '=' signs and
finished by checking that it was now possible to once again consume
composefs-rs from bootc without a bunch of unnecessary version changes
on that side.

Signed-off-by: Allison Karlitskaya <allison.karlitskaya@redhat.com>