fsverity: call sync_data (fdatasync) in test code
This partially fixes #106.

There's a race between the read-write file descriptor being closed,
and the subsequent ioctl to enable fsverify on the read-only copy.
Sometimes the ioctl will return ETXTBSY.  Syncing the data here seems
to be sufficient to resolve this race, but *only* on ext4.  On btrfs,
this is not sufficient.  I *can* fix this on btrfs by doing a
`syncfs()`, but we are trying to avoid doing that here.

Signed-off-by: John Eckersberg <jeckersb@redhat.com>