repository: use OnceCell for objects_dir
Rename our `Repository::object_dir()` accessor to `.objects_dir()` (the
directory is called `"objects"`) and change its implementation to use a
OnceCell.  ostree uses this "initialize on first use" pattern for some
of its directories as well, and I like it.

We use the external `once_cell::` crate (which we already had as a
-devel dependency) because the .try() version of the API is not yet
stable in the standard library.

Clean up our ensure_object() implementation to use `.objects_dir()` and
generally make things a bit more robust:
 - we avoid unnecessary mkdir() calls for directories which almost
   certainly already exist
 - instead of checking merely for the existence of an object file with
   the correct name, we actually measure it now
 - we play around less with joining pathnames (and we can drop a
   now-unused trait helper method on FsVerityHashValue)

Signed-off-by: Allison Karlitskaya <allison.karlitskaya@redhat.com>