src: start using a bit of refcounting
Our `SplitStreamWriter` and `oci::ImageOp` structs contain simple
references to `Repository` which results in some awkward lifetime rules
on those structs.  We can simplify things substantially if we lean into
ref-counting a bit more.

I'm not yet ready to declare that Repository is always refcounted, but
for operations involving splitstreams (including oci downloads) it is
now required.

The ergonomics of this change surprised me.  The Deref trait on `Arc<>`
and the ability to define `self: &Arc<Self>` methods makes this all
quite nice to use.

Signed-off-by: Allison Karlitskaya <allison.karlitskaya@redhat.com>