oci: fix printing of "already" layer IDs
When we already have a layer, we print a message saying as much.  In
fb1de9f482aa ("src/fsverity: up our FsVerityHashValue trait game") we
accidentally changed this to use the debug trait.  This would have been
correct if we were printing a layer ID, but this is an array, so we see
the bytes if we do that.  Revert to the previous behaviour.

Signed-off-by: Allison Karlitskaya <allison.karlitskaya@redhat.com>