src/image.rs: rename to tree.rs
This file has always had a bad name.  Conceptually it represents the
content of an image, but the word "image" has way too many meanings in
this crate.  When I talk to other people about this code I usually find
myself calling it "the in-memory filesystem tree", so let's rename it to
"tree".

This commit was completely automated thanks to LSP tooling and rust-fmt.

Signed-off-by: Allison Karlitskaya <allison.karlitskaya@redhat.com>