src: add new cmdline file
Fix a longstanding TODO in composefs-setup-root for properly dealing
with quoting in cmdline arguments by adding a helper function in a
separate file (we'll use that elsewhere soon, as well).

Also: start using String a bit more.  I think I've been avoiding it as
inefficient but now that I understand more about how it works
(particularly how the length and indices are in byte offsets, not
character offsets), I have a lot less reason to avoid it.

Signed-off-by: Allison Karlitskaya <allison.karlitskaya@redhat.com>