WIP: Cargo.toml: switch to a workspace
Split into a few separate crates:
 - composefs
 - composefs-oci
 - composefs-boot
 - cfsctl
 - composefs-setup-root

This is not a huge improvement in terms of compile speed, and it has
some drawbacks (like 'cargo run' no longer defaulting to cfsctl) but we
might want to do it anyway, for two reasons:

 - I have a couple of experimental pieces of work (like a fuse
   filesystem implementation and an HTTP downloader) that I'd like to
   commit somewhere, but don't want to add to the main composefs crate,
   mostly for dependency reasons.  Features could help there..

 - I've wanted to do xtask stuff for a long time, but it requires
   workspaces...

This is something like the transformation required to make it work.
Maybe let's start discussing it.