examples: workaround dracut xattr handling issues
Newer versions of dracut throw thousands of errors when trying to copy
files into the initramfs in some container setups, and in some cases of
symlinks can fail to include files entirely.  This has led to a
long-lingering problem that makes it difficult to build images locally
(despite them working in CI).

Here's the workaround: DRACUT_NO_XATTR=1.  Thanks to Dick Marinus for
that.

See https://discussion.fedoraproject.org/t/146603/3

This lets us finally increase our Fedora version in the examples to 42.

Signed-off-by: Allison Karlitskaya <allison.karlitskaya@redhat.com>