examples: add a broken example case
Recent changes in the 6.15-rc kernel series have broken
composefs-setup-root when running without using the newest features of
the mount API (ie: with the pre-6.15 feature left enabled, which is the
default).

Add a case to the examples for testing that.  Once it gets working again
it would be useful to make sure it continues to work.

Signed-off-by: Allison Karlitskaya <allison.karlitskaya@redhat.com>